Develop

+++ popos_setup.py
@@ -254,6 +254,17 @@ def main():
     VIMRC_FILE = HOME_DIR + '/.vimrc'
     ETC_SYSFS_CONF_FILE = '/etc/sysfs.conf'
 
+    # Identify Terminal Size
+
+    c = os.popen('stty size', 'r').read().split()[1]
+
+    if int(c) > 120:
+        columns = 120
+    else:
+        columns = int(c)
+
+    # Process CLI arguments
+
     if args.test is True:
         print(args)
         quit()
@@ -272,7 +283,7 @@ def main():
 
     # Create Backup Directory
 
-    print('\n' + '*'*100 + '\n\tCreating Backup Directory for config files...\n' + '*'*100 + '\n')
+    print('\n' + '*'*columns + '\n\tCreating Backup Directory for config files...\n' + '*'*columns + '\n')
 
     if args.backup_directory == 'default':
         backup_directory = HOME_DIR + '/backup/'
@@ -289,27 +300,45 @@ def main():
         print('Backup Directory creation failed.  Exiting script.')
         quit()
 
+    # Identify System Type - unpack os.uname() tuple
+
+    # Read /etc/os-release and decide what OS the system is
+
+    os_info = os.popen('cat /etc/*-release | grep ID', 'r').read()
+
+    if "ubuntu" in os_info:
+        prefix_command = 'sudo apt '
+        prefix_install_command = 'sudo apt install -y '
+    elif 'fedora' in os_info:
+        prefix_command = 'sudo dnf '
+        prefix_install_command = 'sudo dnf install '
+    else:
+        print('OS type not found!')
+        quit()
+
     # Command Variables - Can be user defined.  New variable can be added to update_command array.
 
     pkg_commands = read_json_file('data/pkgcommands.json')
 
     # Read Python Object attributes and store in variables
-    update_command = pkg_commands.get('update_command')
-    full_upgrade_command = pkg_commands.get('full_upgrade_command')
-    favorite_packages = pkg_commands.get('favorite_packages')
-    codec_packages = pkg_commands.get('codec_packages')
-    sshfs_support = pkg_commands.get('sshfs_support')
-    python3_extras = pkg_commands.get('python3_extras')
-    programming_extras = pkg_commands.get('programming_extras')
-    cleanup_packages = pkg_commands.get('cleanup_packages')
-    flatpak_packages = pkg_commands.get('flatpak_packages')
-    snap_packages = pkg_commands.get('snap_packages')
-    set_default_editor = pkg_commands.get('set_default_editor')
+    update_command = prefix_command + str(pkg_commands.get('update_command'))
+    upgrade_command = prefix_command + str(pkg_commands.get('upgrade command'))
+    full_upgrade_command = prefix_command + str(pkg_commands.get('full_upgrade_command'))
+    favorite_packages = prefix_install_command + str(pkg_commands.get('favorite_packages'))
+    codec_packages = prefix_install_command + str(pkg_commands.get('codec_packages'))
+    sshfs_support = prefix_install_command + str(pkg_commands.get('sshfs_support'))
+    python3_extras = prefix_install_command + str(pkg_commands.get('python3_extras'))
+    programming_extras = prefix_install_command + str(pkg_commands.get('programming_extras'))
+    cleanup_packages = prefix_command + str(pkg_commands.get('cleanup_packages'))
+    flatpak_packages = str(pkg_commands.get('flatpak_packages'))
+    # To Do - update these for Fedora
+    snap_packages = str(pkg_commands.get('snap_packages'))
+    set_default_editor = str(pkg_commands.get('set_default_editor'))
 
 
     # Create Array of upgrade commands to parse through
 
-    update_commands_array = [update_command, full_upgrade_command]
+    update_commands_array = [update_command, upgrade_command, full_upgrade_command]
 
     # Create Array of install commands to parse through
 
@@ -328,10 +357,10 @@ def main():
     syntax_commands = read_json_file('data/syntaxcommands.json')
 
     # Read Python Object attributes and store in variables
-    syntax_dot_bash_rc = syntax_commands.get('syntax_dot_bash_rc')
-    syntax_bash_aliases = syntax_commands.get('syntax_bash_aliases')
-    syntax_dot_vimrc = syntax_commands.get('syntax_dot_vimrc')
-    syntax_bluetooth_ERTM_disable = syntax_commands.get('syntax_bluetooth_ERTM_disable')
+    syntax_dot_bash_rc = str(syntax_commands.get('syntax_dot_bash_rc'))
+    syntax_bash_aliases = str(syntax_commands.get('syntax_bash_aliases'))
+    syntax_dot_vimrc = str(syntax_commands.get('syntax_dot_vimrc'))
+    syntax_bluetooth_ERTM_disable = str(syntax_commands.get('syntax_bluetooth_ERTM_disable'))
     
     # Create Array of script variables to parse through
 
@@ -341,7 +370,7 @@ def main():
                             syntax_bluetooth_ERTM_disable
                           ]
 
-    print('\n' + '*'*100 + '\n\tGetting Kernel and other System information...\n' + '*'*100 + '\n')
+    print('\n' + '*'*columns + '\n\tGetting Kernel and other System information...\n' + '*'*columns + '\n')
 
     shell = os.getenv('SHELL')
     kernel = os.system('uname -svr')
@@ -352,25 +381,25 @@ def main():
     skip_items = ' '.join(args.skip)
 
     if "update" in skip_items.lower():
-        print('\n' + '*'*100 + '\n\tSkipping Update Process\n' + '*'*100 + '\n')
+        print('\n' + '*'*columns + '\n\tSkipping Update Process\n' + '*'*columns + '\n')
     else:
-        print('\n' + '*'*100 + '\n\tUpdating System...\n' + '*'*100 + '\n')
+        print('\n' + '*'*columns + '\n\tUpdating System...\n' + '*'*columns + '\n')
 
         process_commands(update_commands_array)
 
     if "install" in skip_items.lower():
-        print('\n' + '*'*100 + '\n\tSkipping Package Install Process\n' + '*'*100 + '\n')
+        print('\n' + '*'*columns + '\n\tSkipping Package Install Process\n' + '*'*columns + '\n')
     else:
-        print('\n' + '*'*100 + '\n\tInstalling Additional Packages...\n' + '*'*100 + '\n')
+        print('\n' + '*'*columns + '\n\tInstalling Additional Packages...\n' + '*'*columns + '\n')
 
         process_commands(install_commamds_array)
 
     # Install 3rd Party Packages
 
     if "3rdparty" in skip_items.lower():
-        print('\n' + '*'*100 + '\n\tSkipping 3rd Party Package Install Process\n' + '*'*100 + '\n')
+        print('\n' + '*'*columns + '\n\tSkipping 3rd Party Package Install Process\n' + '*'*columns + '\n')
     else:
-        print('\n' + '*'*100 + '\n\tInstalling Additional 3rd Party Packages...\n' + '*'*100 + '\n')
+        print('\n' + '*'*columns + '\n\tInstalling Additional 3rd Party Packages...\n' + '*'*columns + '\n')
 
         # Python Foo:  get diretory list and filter for *.deb using Regular Expressions!
         third_party_apps = [val for val in os.listdir(download_dir) if re.search(r'.deb', val)]
@@ -388,13 +417,13 @@ def main():
    
     # Set Default CLI editor...I like Vim!
     
-    print('\n' + '*'*100 + '\n\tSetting the default editor (I like VIM)...\n' + '*'*100 + '\n')
+    print('\n' + '*'*columns + '\n\tSetting the default editor (I like VIM)...\n' + '*'*columns + '\n')
 
     os.system(set_default_editor)
     
     # Add lines to configuration files
 
-    print('\n' + '*'*100 + '\n\tConfiguring the Config files...\n' + '*'*100 + '\n')
+    print('\n' + '*'*columns + '\n\tConfiguring the Config files...\n' + '*'*columns + '\n')
 
     # Create Config File Array to parse through.  Create Dictionary of config files : config script data.
 
@@ -429,13 +458,19 @@ def main():
                 pass
     
     # Post Installation Work
-    print('\n' + '*'*100 + '\n\tPost Installation Cleanup...\n' + '*'*100 + '\n')
+    print('\n' + '*'*columns + '\n\tPost Installation Cleanup...\n' + '*'*columns + '\n')
 
     # System Package Cleanup
 
     print('\nRemoving Unused Packages...\n')
-    os.system('sudo apt autoremove')
-    
+
+    if "ubuntu" in os_info:
+        os.system('sudo apt autoremove')
+    elif 'fedora' in os_info:
+        os.system('sudo dnf clean')
+    else:
+        print('OS type not found!  Not cleaning system.')
+
     # Store a list of installed packages in HOME/backup (overwrite if file exists ">"; NOT append ">>")
 
     print(f'\n Creating a list of installed packages in {backup_directory}...')
@@ -444,9 +479,9 @@ def main():
     # Test to see if reboot is needed
 
     if test_for_file_exists('/var/run/reboot-required'):
-        print('\n' + '*'*100 + '\n\tPlease reboot to complete installation\n' + '*'*100 + '\n')
+        print('\n' + '*'*columns + '\n\tPlease reboot to complete installation\n' + '*'*columns + '\n')
     else:
-        print('\n' + '*'*100 + '\n\tInstallation Complete!\n' + '*'*100 + '\n')
+        print('\n' + '*'*columns + '\n\tInstallation Complete!\n' + '*'*columns + '\n')
 
 if __name__ == '__main__':
